+++ src/test/integration/java/com/example/base/test/util/ParallelTestTimeExecutionListener.java
@@ -1,54 +1,36 @@
-package com.example.base.test.util.stopwatch;
+package com.example.base.test.util;
 
+import com.example.base.test.util.stopwatch.StopWatchUtil;
 import org.springframework.test.context.TestContext;
-import org.springframework.test.context.support.AbstractTestExecutionListener;
 import org.springframework.util.StopWatch;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 
-import java.util.UUID;
-
-public class ParallelTestTimeExecutionListener extends AbstractTestExecutionListener {
+public class ParallelTestTimeExecutionListener extends AbstractTestTimeExecutionListener {
 
     public static final ThreadLocal<StopWatch> threadLocalStopWatch =
             ThreadLocal.withInitial(StopWatch::new);
 
-    private StopWatch stopWatch;
-
-    public static final UUID randomUUID = UUID.randomUUID();
-
-    @Override
-    public void beforeTestClass(TestContext testContext) throws Exception {
-        super.beforeTestClass(testContext);
-        System.out.println("Running test '" + testContext.getTestClass().getSimpleName() + "'...");
-        stopWatch = new StopWatch();
-        stopWatch.start("Total");
-    }
-
-    @Override
-    public void beforeTestMethod(TestContext testContext) throws Exception {
-        super.beforeTestMethod(testContext);
-    }
+    private final SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd-HH-mm-ss");
+    private final Date startedTime = new Date();
 
     @Override
     public void afterTestMethod(TestContext testContext) throws Exception {
+        super.afterTestMethod(testContext);
         StopWatch stopWatch = ParallelTestTimeExecutionListener.threadLocalStopWatch.get();
         StopWatchUtil.exportCsv(stopWatch, exportCsvPathResolver(testContext));
         threadLocalStopWatch.remove();
-        super.afterTestMethod(testContext);
-    }
-
-    @Override
-    public void afterTestClass(TestContext testContext) throws Exception {
-        super.afterTestClass(testContext);
-        stopWatch.stop();
-        System.out.println("The test has been completed" + testContext.getTestClass().getSimpleName());
-        System.out.println(stopWatch.prettyPrint());
     }
 
     private String exportCsvPathResolver(TestContext testContext){
         String packageName = testContext.getTestClass().getPackageName();
         String className = testContext.getTestClass().getSimpleName();
-        String csvFilePath = "./test/reports/" + packageName
-                + "/" + className + "/" + randomUUID + "/" +testContext.getTestMethod()+".csv";
+        String csvFilePath = String.join("/",
+                "./test/reports",
+                packageName,
+                className,
+                dateFormat.format(startedTime),
+                testContext.getTestMethod().getName()) + ".csv";
 
         return csvFilePath;
     }
